Moves static Swarm methods to _ibc.py
